requirements.txt
aiogram==3.15.0
openai==1.54.5
python-dotenv==1.0.1
aiohttp==3.9.5
aiofiles==24.1.0
